+++ core/src/main/java/hzt/sequences/SkipTakeSequence.java
@@ -3,8 +3,8 @@ package hzt.sequences;
 public interface SkipTakeSequence<T> extends Sequence<T> {
 
     @Override
-    Sequence<T> skip(int n);
+    Sequence<T> skip(long n);
 
     @Override
-    Sequence<T> take(int n);
+    Sequence<T> take(long n);
 }
